src/requirements.txt
google-cloud-pubsub
google-cloud-storage
google-cloud-bigquery
google-auth-httplib2
google-auth-oauthlib
google-cloud-secret-manager
google-auth